+++ src/main/java/com/example/Caltizm/Repository/DataRepository.java
@@ -1,10 +1,10 @@
 package com.example.Caltizm.Repository;
 
+import com.example.Caltizm.DTO.BrandDTO;
 import com.example.Caltizm.DTO.CartDTO;
 import com.example.Caltizm.DTO.CategoryDTO;
 import com.example.Caltizm.DTO.ProductDTO;
-import com.example.Caltizm.Service.GetFTADataService;
-import com.example.Caltizm.Service.GetProductDataService;
+import com.example.Caltizm.Service.GetDataService;
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
@@ -12,20 +12,74 @@ import org.springframework.stereotype.Repository;
 import java.util.*;
 
 @Repository
-public class ProductRepository {
+public class DataRepository {
 
     @Autowired
-    GetProductDataService service;
-
-    @Autowired
-    GetFTADataService FTAservice;
+    GetDataService service;
 
     @Autowired
     SqlSession session;
 
+    public void collectAndInsertData(){
+        collectAndInsertBrandData();
+        collectAndInsertProductData();
+    }
+
+    public void collectAndInsertBrandData(){
+        List<BrandDTO> brands = service.collectBrandInfo();
+        // 안전 업데이트 모드 끄기
+        session.update("brand.setSafeUpdateOff");
+
+        // 1. 모든 데이터의 삭제여부를 true로
+        session.update("brand.setDeletedTrue");
+
+        for (BrandDTO brand : brands) {
+            try {
+                // 2. 동일한 이름의 데이터 존재 여부 확인
+                int checkName = session.selectOne("brand.checkName" , brand.getName());
+                if(checkName > 0){
+                    session.update("brand.setDeletedFalse", brand.getName());
+                }else {
+                    session.insert("brand.insert" , brand);
+                    System.out.println("새로운 브랜드 : " + brand.getName());
+                }
+
+            } catch (Exception e) { // 예외 처리
+                System.err.println("Failed to process: " + brand);
+                e.printStackTrace();
+            }
+        }
+
+        // 안전 업데이트 모드 다시 켜기
+        session.update("brand.setSafeUpdateOn");
+
+    }
+
+    public List<BrandDTO> getAllBrand(){
+        List<BrandDTO> brands = session.selectList("brand.selectAll");
+        return brands;
+
+    }
+    public BrandDTO getBrandByName (String name){
+        BrandDTO brand = session.selectOne("brand.selectOne" , name);
+        return brand;
+    }
+    public Map<String , Object> getBrandAndProduct (String name){
+        Map<String , Object> brandAndProduct = new HashMap<>();
+        BrandDTO brand = session.selectOne("brand.selectOne" , name);
+        List<ProductDTO> products = session.selectList("product.selectProductInBrand", name);
+
+        brandAndProduct.put("brand" , brand);
+        brandAndProduct.put("products" , products);
+
+        return brandAndProduct;
+
+    }
+
+
 
     public void collectAndInsertProductData() {
-        Set<ProductDTO> products = service.collectProductDetailsAsync();
+        Set<ProductDTO> products = service.collectProductInfo();
         Set<CategoryDTO> category1 = new HashSet<>();
         for (ProductDTO product : products) {
             CategoryDTO category = new CategoryDTO();
@@ -142,7 +196,7 @@ public class ProductRepository {
         session.update("brand.setSafeUpdateOn");
 
         //FTA 가능 여부
-        Set<String> FTA = FTAservice.collectFTAItemCode();
+        Set<String> FTA = service.collectFTAItemCode();
         for (String product_id : FTA) {
             session.update("product.setFTA", product_id);
         }
@@ -155,7 +209,7 @@ public class ProductRepository {
 
     }
 
-    public ProductDTO getBrandByName(String name) {
+    public ProductDTO getProductByName(String name) {
         ProductDTO product = session.selectOne("product.selectOne", name);
         return product;
     }
@@ -165,7 +219,4 @@ public class ProductRepository {
         CartDTO cartItemInfo = session.selectOne("product.selectCartItemInfo", product_id);
         return cartItemInfo;
     }
-
-
 }
-
