+++ src/main/resources/templates/product/product-detail_test.html
@@ -225,6 +225,7 @@
         <span class="cart-item-quantity" th:text="'수량: ' + ${item.quantity}">수량: 1</span>
         <form th:action="@{/cart/remove}" method="post" style="display:inline;">
             <input type="hidden" name="product_id" th:value="${item.product_id}">
+            <input type="hidden" name="current_product_id" th:value="${product.product_id}">
             <button type="submit" class="remove-item-button">삭제</button>
         </form>
     </div>
